+++ external/js/cothority/tslint.json
